backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.121.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
